+++ docs-src/src/pages/sem/electron-database.tsx
@@ -4,7 +4,7 @@ export default function Page() {
     return Home({
         sem: {
             id: 'gads',
-            title: <>The easiest way to <b className="underline">store</b> and <b className="underline">sync</b> Data in Electron</>,
+            title: <>The easiest way to <b >store</b> and <b >sync</b> Data in Electron</>,
             appName: 'Electron',
             iconUrl: '/files/icons/electron.svg',
             metaTitle: 'Local Electron Database'