+++ docs-src/src/pages/sem/ionic-database.tsx
@@ -6,7 +6,7 @@ export default function Page() {
             id: 'gads',
             metaTitle: 'The local Database for Ionic Apps',
             appName: 'Ionic',
-            title: <>The easiest way to <b className="underline">store</b> and <b className="underline">sync</b> Data in Ionic</>,
+            title: <>The easiest way to <b >store</b> and <b >sync</b> Data in Ionic</>,
             iconUrl: '/files/icons/ionic.svg'
         }
     });