+++ docs-src/src/pages/sem/reddit.tsx
@@ -6,7 +6,7 @@ export default function Page() {
             id: 'reddit',
             metaTitle: 'The local Database for Angular Apps',
             appName: 'Angular',
-            title: <>The easiest way to <b className="underline">store</b> and <b className="underline">sync</b> Data</>
+            title: <>The easiest way to <b >store</b> and <b >sync</b> Data</>
         }
     });
 }